refactor metrics registration and entrypoint
