+++ packages/go/analysis/post/sink.go
@@ -18,38 +18,16 @@ package post
 
 import (
 	"context"
-	"fmt"
 	"log/slog"
 	"runtime"
 	"sync"
 	"time"
 
-	"github.com/prometheus/client_golang/prometheus"
 	"github.com/specterops/bloodhound/packages/go/bhlog/attr"
 	"github.com/specterops/bloodhound/packages/go/graphschema/common"
 	"github.com/specterops/dawgs/graph"
 	"github.com/specterops/dawgs/util/channels"
 )
-
-var (
-	postOperationsVec = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: "analysis",
-		Name:      "post_processing_ops",
-		Help:      "Post-processing operation statistics.",
-	}, []string{
-		"kind",
-		"operation",
-	})
-)
-
-// InitializePostProcessingMetrics registers the analysis post-processing counter with the provided Prometheus registerer.
-func InitializePostProcessingMetrics(registerer prometheus.Registerer) error {
-	if err := registerer.Register(postOperationsVec); err != nil {
-		return fmt.Errorf("failed to register analysis post-processing metrics: %w", err)
-	}
-
-	return nil
-}
 
 func newPropertiesWithFirstSeen() *graph.Properties {
 	newProperties := graph.NewProperties()
